backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
instagrapi>=2.0.0
python-multipart>=0.0.6
pydantic>=2.0.0
python-dotenv>=1.0.0
